Move image logic from create component to item-image component
