+++ src/app/pages/new/create/create.page.html
@@ -35,10 +35,7 @@
   @if(itemFormGroup.valid){
   <app-item-image
     [selectedCategoryItem]="selectedCategoryItem()"
-    [(actionSheetType)]="actionSheetType"
-    [(isActionSheetOpen)]="isActionSheetOpen"
     [(images)]="images"
-    [(selectedImage)]="selectedImage"
   ></app-item-image>
 
   <!-- Voice Notes -->
